requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.34.3

langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33

python-docx==1.2.0
python-pptx==1.0.2
pandas==2.3.2
scikit-learn==1.7.2
numpy==2.2.6
tiktoken==0.11.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.6

python-dotenv==1.0.0
aiofiles==23.2.1
asyncio==3.4.3

beautifulsoup4==4.13.5
pypandoc==1.15
